absorb dummy agent into q_table
